[ADD + REFACTOR]Added listing of winning ballots + major refactor of endpoints correctly subdivided services in each individual macro functionality

+++ app/models/winning_ballots.py
@@ -9,7 +9,7 @@ class WinningBallot(Base):
     lottery_id = Column(Integer, ForeignKey('lotteries.lottery_id'), primary_key=True)
     ballot_id = Column(Integer, ForeignKey('ballots.ballot_id'), nullable=False, unique=True)
     winning_date = Column(Date, nullable=False)
-    winning_amount = Column(Numeric(100, 2), nullable=False)
+    winning_amount = Column(Integer, nullable=False)
 
     lottery = relationship("Lottery", back_populates="winning_ballot_entry")
     ballot = relationship("Ballot", back_populates="winning_entry")
